Added button click events

+++ plugins/core/input/include/plugin_input/mouse.h
@@ -3,8 +3,8 @@
 
 struct mouse_position_t
 {
-    uint32_t x;
-    uint32_t y;
+    double x;
+    double y;
 };
 
 struct mouse_t
@@ -13,6 +13,6 @@ struct mouse_t
     char state;
 };
 
-EVENT_LISTENER2(on_mouse_move, uint32_t x, uint32_t y);
+EVENT_LISTENER2(on_mouse_move, double x, double y);
 EVENT_LISTENER1(on_mouse_button_press, uint32_t btn);
 EVENT_LISTENER1(on_mouse_button_release, uint32_t btn);
